Moved most logic functions form routes folders to logic. Resolved all necessary imports.
