+++ app/controllers/main_routes/__init__.py
@@ -19,7 +19,7 @@ from app.controllers.main_routes import laborHistory
 from app.controllers.main_routes import alterLSF
 from app.controllers.main_routes import studentOverloadApp
 from app.controllers.main_routes import contributors
-from app.controllers.main_routes import download
+from app.logic import download
 from app.controllers.main_routes import laborReleaseForm
 from app.controllers.main_routes import contributors
 from app.controllers.main_routes import studentLaborEvaluation
